requirements.txt
google-ai-generativelanguage==0.6.6
google-api-core==2.19.2
google-api-python-client==2.144.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
pillow==10.4.0
requests==2.32.3
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8